high_entropy_rng/requirements.txt
numpy
matplotlib
pycryptodomex
scipy
